requirements.txt
requirements.txt - Python Dependencies

Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

LLM & AI
langchain==0.1.0
langgraph==0.0.20
openai==1.3.5
httpx==0.25.0

Vector & Embeddings
llama-index==0.9.35
sentence-transformers==2.2.2
pgvector==0.2.4

Database & Cache
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.0
alembic==1.13.0

Authentication
pyjwt==2.8.1
cryptography==41.0.7

Utilities
python-dateutil==2.8.2
requests==2.31.0
pyyaml==6.0.1

Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
pylint==3.0.3
